+++ package.json
@@ -29,7 +29,8 @@
     "axios": "^1.7.9",
     "framer-motion": "^11.15.0",
     "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^7.7.0"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^6.6.3",
@@ -38,6 +39,7 @@
     "@types/node": "^20.17.9",
     "@types/react": "^18.3.14",
     "@types/react-dom": "^18.3.2",
+    "@types/react-router-dom": "^5.3.3",
     "@typescript-eslint/eslint-plugin": "^6.21.0",
     "@typescript-eslint/parser": "^6.21.0",
     "@vitejs/plugin-react-swc": "^3.7.2",
